% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankinma-package.R
\docType{package}
\name{rankinma-package}
\alias{rankinma}
\alias{rankinma-package}
\title{rankinma: Treatment Ranking in Network Meta-Analysis}
\description{
A supportive collection of functions for gathering and ploting treatment ranking metrics after network meta-analysis.
}
\keyword{internal}
