% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_metrics.R
\name{ShowColor}
\alias{ShowColor}
\title{Display color of each treatment}
\usage{
ShowColor(data)
}
\arguments{
\item{data}{DATA of \emph{rankinma} class.}
}
\description{
ShowColor is a function for showing colors of every treatment on plot of
treatment rank metrics in \emph{rankinma}.
}
