% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankinma-package.R
\docType{package}
\name{rankinma-package}
\alias{rankinma}
\alias{rankinma-package}
\title{rankinma: Ranking in Network Meta-Analysis}
\description{
A supportive collection of functions for gathering and plotting treatment ranking metrics after network meta-analysis.
}
\author{
\strong{Maintainer}: Enoch Kang \email{y.enoch.kang@gmail.com} (\href{https://orcid.org/0000-0002-4903-942X}{ORCID}) [translator]

Authors:
\itemize{
  \item Cheihfeng Chen \email{clifchen@tmu.edu.tw} (\href{https://orcid.org/0000-0002-1595-6553}{ORCID})
  \item Wen-Hsuan Hou (\href{https://orcid.org/0000-0002-4376-6298}{ORCID})
  \item Jin-Hua Chen \email{jh_chen@tmu.edu.tw} (\href{https://orcid.org/0000-0002-3130-4125}{ORCID})
  \item Yu-Chieh Chuang \email{b101104079@tmu.edu.tw} (\href{https://orcid.org/0000-0002-7124-6556}{ORCID})
  \item Edwin Chan
}

}
\keyword{internal}
