% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_metrics.R
\name{ShowColor}
\alias{ShowColor}
\title{Display color for each treatment}
\usage{
ShowColor(data)
}
\arguments{
\item{data}{DATA of \emph{rankinma} class.}
}
\value{
\strong{ShowColor()} show a plot of color for each treatment.
}
\description{
ShowColor is a function for showing colors of every treatment on plot of
treatment rank metrics in \emph{rankinma}.
}
\author{
Enoch Kang
}
